Rename heartbeats to cron check-ins
See https://github.com/appsignal/appsignal-elixir/pull/957.

Emit a deprecation warning both when the `heartbeat` helper is used
and when the `Heartbeat` class is initialised, but only once each.
