+++ .changesets/deprecate-heartbeats.md
@@ -1,0 +1,6 @@
+---
+bump: patch
+type: deprecate
+---
+
+Calls to `Appsignal.heartbeat` and to the `Appsignal.Heartbeat` constructor will emit a deprecation warning.
