+++ src/index.ts
@@ -11,4 +11,5 @@ export { expressErrorHandler } from "./instrumentation/express/error_handler"
 export { LoggerLevel, LoggerAttributes } from "./logger"
 export { WinstonTransport } from "./winston_transport"
 export * from "./helpers"
+export * as checkIn from "./check_in"
 export { Heartbeat, heartbeat } from "./heartbeat"
